unify deduction scripts for categories and functors
