+++ databases/catdat/scripts/redundancies.ts
@@ -3,12 +3,12 @@ import {
 	get_categories,
 	get_ignored_redundant_properties,
 	get_normalized_category_implications,
-	get_property_assignments_by_deduction,
 	type NormalizedCategoryImplication,
 } from './utils/categories'
 import {
 	get_deduced_satisfied_properties,
 	get_deduced_unsatisfied_properties,
+	get_property_assignments_by_deduction,
 } from './utils/deduction'
 
 const db = get_client()
@@ -32,7 +32,7 @@ function check_redundant_category_property_assignments() {
 
 	const implications = get_normalized_category_implications(db)
 	const categories = get_categories(db)
-	const assignments = get_property_assignments_by_deduction(db, categories)
+	const assignments = get_property_assignments_by_deduction(db, categories, 'category')
 	const ignore_dict = get_ignored_redundant_properties(db)
 
 	const ignore_count = Object.keys(ignore_dict).reduce(
